fix: resolve cargo fmt and clippy warnings post-v1.3.5
Clears every CI lint gate:

  cargo fmt --all -- --check                          → clean
  cargo clippy --workspace -- -D warnings             → clean
  cargo clippy -p alec-gateway  --features metrics  -- -D warnings → clean
  cargo clippy -p alec-complexity --features gateway -- -D warnings → clean
  cargo clippy --all-targets --all-features -- -D warnings → clean

Fixes applied:

  * `cargo fmt --all` — 23 auto-formatted regions in
    encoder.rs / decoder.rs / protocol.rs /
    context/mod.rs / alec-ffi/src/lib.rs (layout only,
    no semantic change).

  * `doc_overindented_list_items` (≥ clippy 1.94) —
    reflowed every multi-line doc bullet in the Bloc B/C/D
    doc comments so the continuation text sits 2 spaces
    after the `*` instead of column-aligned:
      src/decoder.rs:decode_multi_fixed
      src/encoder.rs:encode_multi_fixed
      alec-ffi/src/lib.rs:alec_heap_init_with_buffer,
        alec_encoder_new_with_config,
        alec_downlink_handler,
        alec_encode_multi_fixed, alec_decode_multi_fixed,
        alec_decoder_export_state,
        alec_decoder_import_state

  * `field_reassign_with_default` — 4 instances
    rewritten to use struct-update syntax:
      alec-ffi/src/lib.rs::AlecEncoderConfig::to_context_config
      src/context/mod.rs (3 evolution tests)

  * `needless_range_loop` — rewrote the post-decode
    observation loop in alec_decode_multi_fixed to use
    `output_slice.iter().take(channel_count).enumerate()`.

  * `casting_to_the_same_type` — removed a
    redundant `max_forward_jump as u16` in
    ctx_version_compatible.

  * `approx_constant` (no_std branch of Pattern::score)
    — replaced the literal `0.693147` with
    `core::f64::consts::LN_2`. Pre-existing, surfaced by
    the bare-metal clippy gate.

  * const-MSRV `div_ceil` — fixed_bitmap_bytes() used
    `usize::div_ceil` which is const-stable only from
    1.73; crate MSRV is 1.70. Replaced with manual
    `(a + 7) / 8`. Pre-existing under stricter clippy.

  * Unused `Result` in 2 benches — bound with `.unwrap()`
    so criterion surfaces a real panic on regressions.

Side-effect: `cbindgen` regenerated include/alec_generated.h
with the new doc-comment layout. No FFI ABI change.

Side-scrub: rustdoc of `alec_downlink_handler` previously
mentioned "Milesight integration" and the "EM500-CO2"
device model. Both were moved to internal docs during the
documentation session but the rustdoc text still leaked.
Replaced with neutral equivalents to match the v1.3.5
public-facing policy.

Tests after fixes: all 19 workspace suites green.
  alec:     158 passed
  alec-ffi: 44 passed, 1 ignored (diag_fixed_sizes)

+++ src/protocol.rs
@@ -302,7 +302,7 @@ pub fn ctx_version_compatible(incoming: u16, last: u16, max_forward_jump: u16) -
     // u16 wrapping subtraction gives the forward distance (0..=65535).
     // Any value <= max_forward_jump is treated as a forward step.
     let forward = incoming.wrapping_sub(last);
-    forward <= max_forward_jump as u16
+    forward <= max_forward_jump
 }
 
 /// Encoding types for data compression
